requirements.txt
google-generativeai
langchain
langchain-google-genai
python-dotenv
fastapi
uvicorn
slowapi
pydantic